Update README.md
So far here is my introduction, more professional by linked in or upwork. This site is supper new for me because I am trying to find new job . Did not mean bad stuff. Just need time to explore all the functions here on Github. Please be so kind to understand. :) Many thanks!